pyproject.toml
[tool.poetry]
name = "auth_system"
version = "0.1.0"
description = "A FastAPI authentication system"
authors = ["ramesh"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.22"
pymysql = "^1.0.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pyotp = "^2.6.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
